F_DOUSU: Force UsU in spite of ARB value
My unlocked US991 reports an ARB of 20, but a UsU flash succeeds and
works properly. Verifying the ARB on most devices is an important
precaution before an unlock however, it would appear I have an odd
device out the bunch.

Leaving this here on the slight chance it's useful to someone in the
future that has a device reporting a suspicious ARB value. I don't
recommend anyone try this as, if you have an ARB higher than 2, you
will likely hard brick your device.